requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0

# Web interface
streamlit>=1.28.0

# Configuration and utilities
PyYAML>=6.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced features
# transformers>=4.30.0
# datasets>=2.14.0
# evaluate>=0.4.0
# scikit-learn>=1.3.0
